feat: compare chart option on single chart
