+++ sections/exchange/TradeCard/Charts/Types/CandlesticksChart.tsx
